refactor: move mocks package to root
