+++ pkg/k8s/pod/client_test.go
@@ -6,7 +6,7 @@ import (
 	"reflect"
 	"testing"
 
-	"github.com/ethersphere/beekeeper/pkg/k8s/mocks"
+	mock "github.com/ethersphere/beekeeper/mocks/k8s"
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes"
@@ -48,13 +48,13 @@ func TestSet(t *testing.T) {
 		{
 			name:      "create_error",
 			podName:   "create_bad",
-			clientset: mocks.NewClientset(),
+			clientset: mock.NewClientset(),
 			errorMsg:  fmt.Errorf("creating pod create_bad in namespace test: mock error: cannot create pod"),
 		},
 		{
 			name:      "update_error",
 			podName:   "update_bad",
-			clientset: mocks.NewClientset(),
+			clientset: mock.NewClientset(),
 			errorMsg:  fmt.Errorf("updating pod update_bad in namespace test: mock error: cannot update pod"),
 		},
 	}
@@ -130,7 +130,7 @@ func TestDelete(t *testing.T) {
 		{
 			name:      "delete_error",
 			podName:   "delete_bad",
-			clientset: mocks.NewClientset(),
+			clientset: mock.NewClientset(),
 			errorMsg:  fmt.Errorf("deleting pod delete_bad in namespace test: mock error: cannot delete pod"),
 		},
 	}
