+++ pkg/k8s/service/client_test.go
@@ -6,7 +6,7 @@ import (
 	"reflect"
 	"testing"
 
-	"github.com/ethersphere/beekeeper/pkg/k8s/mocks"
+	mock "github.com/ethersphere/beekeeper/mocks/k8s"
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes"
@@ -65,19 +65,19 @@ func TestSet(t *testing.T) {
 		{
 			name:        "create_error",
 			serviceName: "create_bad",
-			clientset:   mocks.NewClientset(),
+			clientset:   mock.NewClientset(),
 			errorMsg:    fmt.Errorf("creating service create_bad in namespace test: mock error: cannot create service"),
 		},
 		{
 			name:        "update_error",
 			serviceName: "update_bad",
-			clientset:   mocks.NewClientset(),
+			clientset:   mock.NewClientset(),
 			errorMsg:    fmt.Errorf("updating service update_bad in namespace test: mock error: cannot update service"),
 		},
 		{
 			name:        "get_error",
 			serviceName: "get_bad",
-			clientset:   mocks.NewClientset(),
+			clientset:   mock.NewClientset(),
 			errorMsg:    fmt.Errorf("getting service get_bad in namespace test: mock error: unknown"),
 		},
 	}
@@ -153,7 +153,7 @@ func TestDelete(t *testing.T) {
 		{
 			name:        "delete_error",
 			serviceName: "delete_bad",
-			clientset:   mocks.NewClientset(),
+			clientset:   mock.NewClientset(),
 			errorMsg:    fmt.Errorf("deleting service delete_bad in namespace test: mock error: cannot delete service"),
 		},
 	}
