+++ pkg/k8s/serviceaccount/serviceaccount_test.go
@@ -6,7 +6,7 @@ import (
 	"reflect"
 	"testing"
 
-	"github.com/ethersphere/beekeeper/pkg/k8s/mocks"
+	mock "github.com/ethersphere/beekeeper/mocks/k8s"
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes"
@@ -55,13 +55,13 @@ func TestSet(t *testing.T) {
 		{
 			name:       "create_error",
 			secretName: "create_bad",
-			clientset:  mocks.NewClientset(),
+			clientset:  mock.NewClientset(),
 			errorMsg:   fmt.Errorf("creating service account create_bad in namespace test: mock error: cannot create service account"),
 		},
 		{
 			name:       "update_error",
 			secretName: "update_bad",
-			clientset:  mocks.NewClientset(),
+			clientset:  mock.NewClientset(),
 			errorMsg:   fmt.Errorf("updating service account update_bad in namespace test: mock error: cannot update service account"),
 		},
 	}
@@ -149,7 +149,7 @@ func TestDelete(t *testing.T) {
 		{
 			name:       "delete_error",
 			secretName: "delete_bad",
-			clientset:  mocks.NewClientset(),
+			clientset:  mock.NewClientset(),
 			errorMsg:   fmt.Errorf("deleting service account delete_bad in namespace test: mock error: cannot delete service account"),
 		},
 	}
